pyproject.toml
[project]
name = "rag-chatbot"
version = "0.1.0"
description = "rag chatbot"
authors = "Dang Anh Quan"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "openai-agents>=0.2.4",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "qdrant-client~=1.13.1",
    "uvicorn>=0.35.0",
]

